refactored mask_gen to reflect actual repo standard
