+++ assets/directory_structure.md
