+++ assets/inspect_diagram.jpg
