--- conf/unet_conf/default.yaml
@@ -1,3 +1,0 @@
-learning_rate: 0.0001
-batch_size: 32
-epochs: 30
