+++ environment.yaml
@@ -1,20 +1,21 @@
-name: field_segmentation_training
+name: mask_generation_env
 channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.10
-  - pip
+  - python=3.11
+  - opencv
   - numpy
-  - pandas
-  - matplotlib
-  - pillow 
+  - scikit-image
+  - pillow
+  - pathlib
   - pytorch
   - torchvision
-  - tqdm
-  - pyyaml
-  - ultralytics  
-  
+  - pandas
+  - hydra-core
+  - omegaconf
+  - fiftyone
+  - pip
   - pip:
-      - torchmetrics
-      - omegaconf
+      - ultralytics
+      - segmentation-models-pytorch
