--- mask_generation/conf/config.yaml
@@ -1,38 +1,0 @@
-# Hydra defaults and settings. 
-mode: ??? # Options: mask_gen
-
-project:
-  name: test
-
-defaults:
-  - paths: default
-  - mask_refine: default
-  - inspect: default
-  - _self_
-
-hydra:
-  run:
-    dir: ${paths.logdir}/${job.job_now}
-   
-
-tasks:
-  sync: false # skip for now; copies the most up-to-date version of the files needed to run the mask_gen mode from LTS to here
-  detect_weeds: false
-  unet_segment_weeds: false
-  mask_inspect: false
-  mask_refine: true
-
-test:
-  enabled: true
-
-mask_inspect:
-  save2lts: false
-
-general:
-  task: # Leave empty as placeholder
-
-job:
-  job_now_date: &nowdir ${now:%Y-%m-%d}
-  job_now_time: ${now:%H_%M_%S}
-  job_now: ${job.job_now_date}/${job.job_now_time} 
-  jobdir: ${paths.datadir}/${job.job_now}
