--- mask_generation/conf/hydra/job_logging/custom.yaml
@@ -1,18 +1,0 @@
-version: 1
-formatters:
-  simple:
-    format: '[%(asctime)s][%(name)s][%(levelname)s] - %(message)s'
-handlers:
-  console:
-    class: logging.StreamHandler
-    formatter: simple
-    stream: ext://sys.stdout
-  file:
-    class: logging.FileHandler
-    formatter: simple
-    filename: ${hydra.runtime.output_dir}/FIELD_REPORT.log
-root:
-  handlers:
-    - console
-    - file
-disable_existing_loggers: false
