--- mask_generation/conf/inspect/default.yaml
@@ -1,3 +1,0 @@
-dataset_name: testing_inspect_maskgen
-port: 5050
-
