--- mask_generation/projects/.gitignore
@@ -1,7 +1,0 @@
-# Ignore everything
-*
-!.gitignore
-# Except the test directory and its contents
-!test/
-!test/developed-images/
-!test/developed-images/**
