+++ src/utils/copy_images_from_lts.py
